# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API endpoint documentation for the ChemistKe api project.",
        "title": "ChemistKe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ChemistKe Support",
            "url": "https://chemist.co.ke/support",
            "email": "Support@chemist.co.ke"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Get products based on page and limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PaginationRequest Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PaginationRequest Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Creates a new Product",
                "parameters": [
                    {
                        "description": "Product information to create",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update a given product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update a  product",
                "parameters": [
                    {
                        "description": "Product information to create",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        },
        "/products/get-products-by-brand": {
            "get": {
                "description": "Get products based on a given brand name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products based on brand",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductReq Brand",
                        "name": "brand",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        },
        "/products/get-products-by-category": {
            "get": {
                "description": "Get products based on a givne Category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products Based on a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductReq Category",
                        "name": "category",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        },
        "/products/get-products-by-sub-category": {
            "get": {
                "description": "Get products based on a given sub-category, page, and limit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get products based on a sub-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductReq SubCategory",
                        "name": "sub-category",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ProductReq Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        },
        "/products/images/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an image for a given product using multipart/form-data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Upload a product image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "product-id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Type of the image (e.g. thumbnail, banner, general)",
                        "name": "image-type",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully uploaded image",
                        "schema": {
                            "$ref": "#/definitions/product_proto.UploadProdctImagesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.ErrResponse"
                        }
                    }
                }
            }
        },
        "/products/images/{id}": {
            "get": {
                "description": "Get the images of a given product id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "a product's images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/routes.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "product_proto.UploadProdctImagesResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "routes.ErrResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "routes.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "routes.Product": {
            "type": "object",
            "required": [
                "brand",
                "category",
                "description",
                "name",
                "price",
                "quantity",
                "sub-category"
            ],
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "J\u0026J"
                },
                "category": {
                    "type": "string",
                    "example": "Antibiotics"
                },
                "description": {
                    "type": "string",
                    "example": "product description"
                },
                "id": {
                    "type": "string",
                    "example": "f183e73c-687d-44ad-83e6-636ecbb7a7d8"
                },
                "name": {
                    "type": "string",
                    "example": "Amoxilin"
                },
                "price": {
                    "type": "number",
                    "example": 300
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                },
                "sub-category": {
                    "type": "string",
                    "example": "Mild"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}